refactor(WIP): restructuring to include base widget
