+++ stac_ipyleaflet/__init__.py
@@ -4,4 +4,4 @@ __author__ = """Aimee Barciauskas"""
 __email__ = 'aimee@developmentseed.org'
 __version__ = '0.1.0'
 
-from .core import *
+from .base_widget import *
